backend/requirements.txt
alembic==1.18.4
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.84.0
anyio==4.12.1
bcrypt==5.0.0
billiard==4.2.4
boto3==1.42.66
botocore==1.42.66
celery==5.6.2
certifi==2026.2.25
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
distro==1.9.0
docstring_parser==0.17.0
ecdsa==0.19.1
faiss-cpu==1.13.2
fastapi==0.135.1
filelock==3.25.2
fsspec==2026.2.0
h11==0.16.0
hf-xet==1.4.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.6.0
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
jiter==0.13.0
jmespath==1.1.0
joblib==1.5.3
kombu==5.6.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.3
openai==2.26.0
packaging==26.0
pandas==3.0.1
passlib==1.7.4
pillow==12.1.1
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pyasn1==0.6.2
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.2
pytest-asyncio==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.22
PyYAML==6.0.3
redis==7.3.0
regex==2026.2.28
replicate==1.0.7
rich==14.3.3
rsa==4.9.1
s3transfer==0.16.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.2.3
setuptools==82.0.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.48
starlette==0.52.1
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.10.0
tqdm==4.67.3
transformers==5.3.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.41.0
vine==5.1.0
wcwidth==0.6.0
